refactor(review): hoist offline result helpers off orchestration
Stop the agent module importing private packet/failure helpers from
offline_review, and parse unified-diff b/ paths through one helper.
