+++ src/mergecraft/offline_review.py
@@ -2,15 +2,11 @@
 
 from __future__ import annotations
 
-import asyncio
 import os
 import subprocess
 import tempfile
-from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
-
-from loguru import logger
 
 from mergecraft.analyzers.finding import (
     STRUCTURED_OUTPUT_REQUIRED_MSG,
@@ -24,13 +20,21 @@ from mergecraft.analyzers.trust import (
     derive_source_trust_tier,
 )
 from mergecraft.config import load_repo_settings
+from mergecraft.review.offline_agent import run_offline_agent_review
+from mergecraft.review.offline_result import (
+    OfflineReviewResult as OfflineReviewResult,
+)
+from mergecraft.review.offline_result import (
+    _offline_error_outcome,
+    _offline_failure,
+)
+from mergecraft.review.offline_stages import changed_paths_from_unified_diff
 from mergecraft.run_outcome import RunOutcome
 from mergecraft.utils.agent_resolve import resolve_model
 from mergecraft.utils.fence import Fence, render_untrusted
 from mergecraft.utils.offline_diff import DiffMaterialization, summarize_diff
 from mergecraft.utils.run_bounds import (
     BudgetExhausted,
-    RunBounds,
     ScopeReduction,
     apply_diff_line_budget,
     budget_exhaustion_outcome,
@@ -48,53 +52,9 @@ if TYPE_CHECKING:
     from collections.abc import Callable
 
     from mergecraft.config.settings import CliTrustOverride
-    from mergecraft.mcp.context import ToolContext
     from mergecraft.review.engine import ReviewEngine
     from mergecraft.tracing.review_context import ReviewContext
     from mergecraft.utils.source_resolve import ResolvedWorkspace
-
-
-@dataclass(slots=True)
-class OfflineReviewResult:
-    success: bool
-    output: str | None = None
-    error: str | None = None
-    diff_path: str | None = None
-    empty_diff: bool = False
-    structured_output: str | None = None
-    # On-disk path of the run's merge evidence packet (#47 / #96), or None
-    # when no packet was produced (dry run, empty diff, emission failure).
-    evidence_packet_path: str | None = None
-    outcome: RunOutcome | None = None
-    scope_reduction: object | None = None
-
-
-def _offline_failure(
-    *,
-    error: str,
-    outcome: RunOutcome,
-    diff_path: str | None = None,
-    evidence_packet_path: str | None = None,
-    output: str | None = None,
-    structured_output: str | None = None,
-) -> OfflineReviewResult:
-    return OfflineReviewResult(
-        success=False,
-        error=error,
-        output=output,
-        structured_output=structured_output,
-        diff_path=diff_path,
-        evidence_packet_path=evidence_packet_path,
-        outcome=outcome,
-    )
-
-
-def _offline_error_outcome(exc: BaseException) -> RunOutcome:
-    if isinstance(exc, (TimeoutError, asyncio.TimeoutError, subprocess.TimeoutExpired)):
-        return RunOutcome.timed_out
-    if isinstance(exc, ValueError):
-        return RunOutcome.configuration_error
-    return RunOutcome.infra_error
 
 
 def parse_offline_review_findings(result: OfflineReviewResult) -> list[Finding]:
@@ -172,8 +132,8 @@ def _filter_diff_to_paths(diff_text: str, kept_paths: list[str]) -> str:
         if line.startswith("diff --git "):
             if current and current_path in keep:
                 blocks.append("".join(current))
-            parts = line.split()
-            current_path = parts[3].removeprefix("b/") if len(parts) >= 4 else None
+            parsed = changed_paths_from_unified_diff(line)
+            current_path = parsed[0] if parsed else None
             current = [line]
             continue
         if current:
@@ -263,66 +223,6 @@ def _finish_offline_result(
 
         store_review_result(cache_key, finished)
     return finished
-
-
-def _offline_change_id(cwd: Path, materialization: DiffMaterialization) -> str:
-    """Return the ``change_id`` an offline review attests to.
-
-    There is no pull request, so the packet addresses the local working tree
-    and the base it was diffed against — enough for a human to reconstruct
-    what was reviewed.
-    """
-    base = materialization.base_ref or "patch"
-    return f"local/{cwd.name}@{base}"
-
-
-def _emit_offline_packet(
-    tool_context: ToolContext,
-    *,
-    cwd: Path,
-    materialization: DiffMaterialization,
-    run_succeeded: bool,
-    structured_output: str | None,
-    output_path: Path | None,
-) -> str | None:
-    """Emit the evidence packet for an offline review (#96).
-
-    The offline path holds the agent's findings in typed form (its
-    ``set_output`` payload), so they are merged into the packet on top of the
-    analyzer findings. A malformed payload is skipped rather than fatal — the
-    caller reports that error separately, and a packet with analyzer evidence
-    only still beats no packet.
-    """
-    from mergecraft.evidence.run_packet import emit_run_packet
-
-    extra: list[Finding] = []
-    if structured_output:
-        try:
-            # ``parse_findings_payload`` validates and then dumps back to dicts;
-            # the packet dedupes on ``Finding.fingerprint``, so re-type them.
-            extra = [
-                Finding.model_validate(row) for row in parse_findings_payload(structured_output)
-            ]
-        except ValueError as exc:
-            logger.debug("offline evidence packet: unparsable structured output — {}", exc)
-
-    written = emit_run_packet(
-        tool_context,
-        run_succeeded=run_succeeded,
-        change_id=_offline_change_id(cwd, materialization),
-        extra_findings=extra,
-        output_path=output_path,
-    )
-    from mergecraft.evidence.run_manifest import build_run_manifest
-
-    manifest = build_run_manifest(
-        cwd=cwd,
-        model=tool_context.resolved_model or tool_context.tool_state.model or "(unresolved)",
-        agent_id=tool_context.agent_id,
-        prompt_text=materialization.path.read_text(encoding="utf-8"),
-    )
-    logger.info("» run manifest fingerprints: {}", manifest)
-    return str(written) if written else None
 
 
 def _apply_tracing_cli_overrides(
@@ -568,11 +468,7 @@ async def _run_offline_diff_review(
         )
         if trust_tier == "untrusted":
             diff_text = built.path.read_text(encoding="utf-8")
-            paths = [
-                line.split()[3].removeprefix("b/")
-                for line in diff_text.splitlines()
-                if line.startswith("diff --git ") and len(line.split()) >= 4
-            ]
+            paths = changed_paths_from_unified_diff(diff_text)
             kept = filter_confined_paths(cwd, paths)
             if kept != paths:
                 filtered = _filter_diff_to_paths(diff_text, kept)
@@ -671,7 +567,7 @@ async def _run_offline_diff_review(
                 from_cache = True
                 return cached
 
-        return await _run_agent_review(
+        return await run_offline_agent_review(
             cwd=cwd,
             materialization=materialization,
             prompt=prompt,
@@ -740,33 +636,3 @@ async def _run_offline_diff_review(
                 os.environ.pop(key, None)
             else:
                 os.environ[key] = value
-
-
-async def _run_agent_review(
-    *,
-    cwd: Path,
-    materialization: DiffMaterialization,
-    prompt: str,
-    model: str | None,
-    tmpdir: Path,
-    output_schema: dict[str, Any] | None = None,
-    evidence_packet_path: Path | None = None,
-    trust_tier: str = "trusted",
-    run_bounds: RunBounds | None = None,
-    on_finding: Callable[[dict[str, Any]], None] | None = None,
-) -> OfflineReviewResult:
-    """Dispatch the Review agent; implementation lives in ``review.offline_agent``."""
-    from mergecraft.review.offline_agent import run_offline_agent_review
-
-    return await run_offline_agent_review(
-        cwd=cwd,
-        materialization=materialization,
-        prompt=prompt,
-        model=model,
-        tmpdir=tmpdir,
-        output_schema=output_schema,
-        evidence_packet_path=evidence_packet_path,
-        trust_tier=trust_tier,
-        run_bounds=run_bounds,
-        on_finding=on_finding,
-    )
